pyproject.toml
[project]
name = "demonstrated-feedback"
version = "0.1.0"
description = "Show, Don't Tell: Aligning Language Models with Demonstrated Feedback"
readme = "README.md"
requires-python = ">=3.10.9"
dependencies = [
    "accelerate==0.30.1",
    "alignment-handbook",
    "bitsandbytes==0.43.1",
    "datasets==2.19.1",
    "deepspeed==0.12.2",
    "einops==0.8.0",
    "evaluate==0.4.0",
    "hf-doc-builder>=0.4.0",
    "hf_transfer>=0.1.4",
    "huggingface-hub==0.23.1",
    "ninja>=1.11.1",
    "numpy==1.24.2",
    "packaging>=23.0",
    "parameterized>=0.9.0",
    "peft==0.9.0",
    "protobuf<=3.20.2",
    "safetensors>=0.3.3",
    "sentencepiece>=0.1.99",
    "scipy",
    "tensorboard",
    "torch==2.1.2",
    "transformers==4.41.1",
    "trl==0.8.6",
    "jinja2>=3.0.0",
    "tqdm>=4.64.1",
    "rich",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "black==23.1.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "pytest",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu121" },
]
alignment-handbook = { git = "https://github.com/huggingface/alignment-handbook.git", rev = "606d2e954fd17999af40e6fb4f712055ca11b2f0" }

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.workspace]
members = ["alignment-handbook"]
